requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
frozenlist==1.4.0
idna==3.4
isort==5.12.0
magic-filter==1.0.11
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pyaes==1.6.1
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
rsa==4.9
Telethon==1.30.3
typing_extensions==4.7.1
yarl==1.9.2
